requirements.txt
# local package
-e .

# external requirements
python-dotenv>=0.5.1
click
rich
python-box
PyYAML
ruffus